Fixed format of state message
* Formatting of CValidationState reject message now happens when an error
occurs in the RPC call sendrawtransaction. Previously, some debug output
was lost and the reasoning for the failure of a transaction rejected
from the mempool was ambiguous. Also, there's no reason to not use the
FormatStateMessage function here, it already does the work of
concatenating the reject reason string with an error code.